Creation of model MenuItem
The model and files related to it was created using the command "rails generate model MenuItem name:string, description:string, price:decimal{4,2}"

The model was configurate with "validates" helper and the reference "belongs_to" to model Menu.
Model menu was updated with reference to the MenuItems model, with a "has_many" reference.

Created and configured migration for "menu_items", which was configured with a foreign key to "menu" table.
The database schema was updated.