Multiple menus
Added associations between models Menu and Restaurant in the "Restaurant" and "Menu" files. Restaurants can have one or more menus and a Menu belongs to a restaurant.
Modified uniqueness constraint in "Menu" model for only scope of Restaurant, which allow diferents restaurants have same name of menu, but unique for one restaurant.

In the table "menus" I create an integer field "restaurant_id" which allowed associate a restaurant with a menu. After this, if table "menus" have records, that table must to be backpopulate
with "restaurant_id" association, which will depend of the business logic.

Once the column "restaurant_id" is backpopulate with verified "restaurant ids" the column can have next constraints: not null, foreign_key, unique.
This constraints was added in a new migration "AddForeignKeyRestaurantToMenu"

The database schema was updated to reflect this changes.