MenuItem update
MenuItem name is not duplicated for each menu.
In model "MenuItem is added uniqueness validation with "menu" scope.
Constraints for this uniquenesses added by a migration, which ensure menu is not empty and is unique with the name.
The database schema is updated to reflect this changes
Updated unit test for "MenuItem" model.