Populate script
The model "MenuItem" is updated in order to import the "restaurant_data.json" file and populate the menu system.
An migration is created for "MenuItem" column "description" was removed from the not null constraint.
The "restaurant_data.json" is created in a folder for scripts created in app/scripts/populate_menu, with the script
which populate the database.
The database schema is updated.

The Populate script

Use ActiveRecord::Base.transaction for not persiste data if there is any error while executing the script, rolling back in the database.
A "dry_run" option is evaluated for run a dry run or a real one.
The script is verbose for easy identify the possible errors happen during population in database.
Two arrays are created, one with successful restaurant id, for successfull execution for each restaurant, and one with the names of Restaurants
if any error happen, for easy identify and debug.